# packages/contracts/swagger/spotify.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully logged in",
            "headers": {
              "Set-Cookie": {
                "description": "HttpOnly cookies: access_token и refresh_token",
                "schema": {
                  "type": "string",
                  "example": "access_token=<jwt>; HttpOnly; Path=/; SameSite=Lax;\nrefresh_token=<jwt>; HttpOnly; Path=/; SameSite=Lax;"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "example": {
                  "errors": [
                    {
                      "field": "email",
                      "message": "email must be an email"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          },
          "405": {
            "description": "Method not allowed"
          },
          "408": {
            "description": "Request timeout"
          },
          "429": {
            "description": "Too many requests"
          },
          "500": {
            "description": "Internal server error"
          },
          "501": {
            "description": "Not implemented"
          },
          "502": {
            "description": "Bad gateway"
          },
          "503": {
            "description": "Service unavailable"
          },
          "504": {
            "description": "Gateway timeout"
          },
          "505": {
            "description": "HTTP version not supported"
          },
          "507": {
            "description": "Insufficient storage"
          },
          "508": {
            "description": "Loop detected"
          },
          "default": {
            "description": ""
          }
        },
        "summary": "User login",
        "tags": ["Auth"]
      }
    },
    "/auth/registration": {
      "post": {
        "operationId": "AuthController_registration",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully registered"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "example": {
                  "errors": [
                    {
                      "field": "email",
                      "message": "email must be an email"
                    }
                  ]
                }
              }
            }
          },
          "405": {
            "description": "Method not allowed"
          },
          "408": {
            "description": "Request timeout"
          },
          "409": {
            "description": "User already exists"
          },
          "429": {
            "description": "Too many requests"
          },
          "500": {
            "description": "Internal server error"
          },
          "501": {
            "description": "Not implemented"
          },
          "502": {
            "description": "Bad gateway"
          },
          "503": {
            "description": "Service unavailable"
          },
          "504": {
            "description": "Gateway timeout"
          },
          "505": {
            "description": "HTTP version not supported"
          },
          "507": {
            "description": "Insufficient storage"
          },
          "508": {
            "description": "Loop detected"
          },
          "default": {
            "description": ""
          }
        },
        "summary": "User registration",
        "tags": ["Auth"]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Successfully logged out"
          },
          "401": {
            "description": "Unauthorized"
          },
          "405": {
            "description": "Method not allowed"
          },
          "408": {
            "description": "Request timeout"
          },
          "429": {
            "description": "Too many requests"
          },
          "500": {
            "description": "Internal server error"
          },
          "501": {
            "description": "Not implemented"
          },
          "502": {
            "description": "Bad gateway"
          },
          "503": {
            "description": "Service unavailable"
          },
          "504": {
            "description": "Gateway timeout"
          },
          "505": {
            "description": "HTTP version not supported"
          },
          "507": {
            "description": "Insufficient storage"
          },
          "508": {
            "description": "Loop detected"
          },
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "User logout",
        "tags": ["Auth"]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Token refreshed",
            "headers": {
              "Set-Cookie": {
                "description": "HttpOnly cookies: access_token",
                "schema": {
                  "type": "string",
                  "example": "access_token=<jwt>; HttpOnly; Path=/; SameSite=Lax;"
                }
              }
            }
          },
          "400": {
            "description": "Refresh token not provided"
          },
          "401": {
            "description": "Invalid or expired refresh token"
          },
          "405": {
            "description": "Method not allowed"
          },
          "408": {
            "description": "Request timeout"
          },
          "429": {
            "description": "Too many requests"
          },
          "500": {
            "description": "Internal server error"
          },
          "501": {
            "description": "Not implemented"
          },
          "502": {
            "description": "Bad gateway"
          },
          "503": {
            "description": "Service unavailable"
          },
          "504": {
            "description": "Gateway timeout"
          },
          "505": {
            "description": "HTTP version not supported"
          },
          "507": {
            "description": "Insufficient storage"
          },
          "508": {
            "description": "Loop detected"
          },
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "refresh_token": []
          }
        ],
        "summary": "Refresh access token",
        "tags": ["Auth"]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully logged out",
            "content": {
              "application/json": {
                "example": {
                  "createdAt": "2023-10-01T12:00:00.000Z",
                  "email": "user@example.com",
                  "id": "1234567890abcdef",
                  "updatedAt": "2023-10-01T12:00:00.000Z",
                  "username": "user123",
                  "avatar": null,
                  "description": "Some description about the user"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "405": {
            "description": "Method not allowed"
          },
          "408": {
            "description": "Request timeout"
          },
          "429": {
            "description": "Too many requests"
          },
          "500": {
            "description": "Internal server error"
          },
          "501": {
            "description": "Not implemented"
          },
          "502": {
            "description": "Bad gateway"
          },
          "503": {
            "description": "Service unavailable"
          },
          "504": {
            "description": "Gateway timeout"
          },
          "505": {
            "description": "HTTP version not supported"
          },
          "507": {
            "description": "Insufficient storage"
          },
          "508": {
            "description": "Loop detected"
          },
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Get current authenticated user",
        "tags": ["Auth"]
      }
    },
...
    "/playlists": {
      "get": {
        "operationId": "PlaylistsController_getAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "path",
            "description": "Page number for pagination",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "path",
            "description": "Number of items per page",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Method not allowed"
          },
          "408": {
            "description": "Request timeout"
          },
          "429": {
            "description": "Too many requests"
          },
          "500": {
            "description": "Internal server error"
          },
          "501": {
            "description": "Not implemented"
          },
          "502": {
            "description": "Bad gateway"
          },
          "503": {
            "description": "Service unavailable"
          },
          "504": {
            "description": "Gateway timeout"
          },
          "505": {
            "description": "HTTP version not supported"
          },
          "507": {
            "description": "Insufficient storage"
          },
          "508": {
            "description": "Loop detected"
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all playlists with pagination and filters",
        "tags": ["Playlists"]
      },
      "post": {
        "operationId": "PlaylistsController_post",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlaylistDto"
              }
            }
          }
        },
        "responses": {
          "405": {
            "description": "Method not allowed"
          },
          "408": {
            "description": "Request timeout"
          },
          "429": {
            "description": "Too many requests"
          },
          "500": {
            "description": "Internal server error"
          },
          "501": {
            "description": "Not implemented"
          },
          "502": {
            "description": "Bad gateway"
          },
          "503": {
            "description": "Service unavailable"
          },
          "504": {
            "description": "Gateway timeout"
          },
          "505": {
            "description": "HTTP version not supported"
          },
          "507": {
            "description": "Insufficient storage"
          },
          "508": {
            "description": "Loop detected"
          },
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Create a new playlist",
        "tags": ["Playlists"]
      }
    },
    "/playlists/{id}": {
      "get": {
        "operationId": "PlaylistsController_getById",
        "parameters": [],
        "responses": {
          "405": {
            "description": "Method not allowed"
          },
          "408": {
            "description": "Request timeout"
          },
          "429": {
            "description": "Too many requests"
          },
          "500": {
            "description": "Internal server error"
          },
          "501": {
            "description": "Not implemented"
          },
          "502": {
            "description": "Bad gateway"
          },
          "503": {
            "description": "Service unavailable"
          },
          "504": {
            "description": "Gateway timeout"
          },
          "505": {
            "description": "HTTP version not supported"
          },
          "507": {
            "description": "Insufficient storage"
          },
          "508": {
            "description": "Loop detected"
          },
          "default": {
            "description": ""
          }
        },
        "summary": "Get playlist by id",
        "tags": ["Playlists"]
      },
      "put": {
        "operationId": "PlaylistsController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlaylistDto"
              }
            }
          }
        },
        "responses": {
          "405": {
            "description": "Method not allowed"
          },
          "408": {
            "description": "Request timeout"
          },
          "429": {
            "description": "Too many requests"
          },
          "500": {
            "description": "Internal server error"
          },
          "501": {
            "description": "Not implemented"
          },
          "502": {
            "description": "Bad gateway"
          },
          "503": {
            "description": "Service unavailable"
          },
          "504": {
            "description": "Gateway timeout"
          },
          "505": {
            "description": "HTTP version not supported"
          },
          "507": {
            "description": "Insufficient storage"
          },
          "508": {
            "description": "Loop detected"
          },
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Update playlist by id",
        "tags": ["Playlists"]
      }
    },
...
  },
  "info": {
    "title": "@spotify/api",
    "description": "@spotify/api Swagger documentation",
    "version": "0.0.1",
    "contact": {
      "name": "Lordpluha",
      "url": "https://github.com/Lordpluha",
      "email": "vladislavteslyukofficial@gmail.com"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    },
    {
      "url": "https://spotify-clone-api-jp5z.onrender.com/",
      "description": "Remote dev server"
    }
  ],
  "components": {
    "securitySchemes": {
      "oauth2": {
        "type": "openIdConnect",
        "flows": {}
      }
    },
    "schemas": {
...
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "password123"
          }
        },
        "required": ["email", "password"]
      },
      "RegistrationDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "New user email",
            "example": "newuser@example.com"
          },
          "password": {
            "type": "string",
            "description": "New user password",
            "example": "password123"
          }
        },
        "required": ["email", "password"]
      },
...
      "CreatePlaylistDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Playlist title"
          },
          "description": {
            "type": "string",
            "description": "",
            "example": "user123"
          }
        },
        "required": ["title"]
      },
      "UpdatePlaylistDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Playlist title"
          },
          "description": {
            "type": "string",
            "description": "",
            "example": "user123"
          }
        },
        "required": ["title"]
      },
...
    }
  },
...
}